+++ src/Resources/CursorCollection.php
@@ -87,6 +87,6 @@ abstract class CursorCollection extends ResourceCollection
                     ? $page->previous()
                     : $page->next();
             }
-        }))->setResponse($this->response);
+        }))->setOrigin($this->getOrigin());
     }
 }